requirements.txt
gspread
google-auth
pandas
jinja2
pdfkit
matplotlib
